Move curve widgets to widgets/curves/
