+++ rtgui/shcselector.cc
@@ -22,8 +22,8 @@
 #include "shcselector.h"
 
 #include "multilangmgr.h"
-#include "mycurve.h"
 #include "rtscalable.h"
+#include "widgets/curves/mycurve.h"
 
 SHCSelector::SHCSelector() : movingPosition(-1), tmpX(0.0), tmpPos(0.0), wslider(0.0), cl(nullptr), coloredBar(RTO_Left2Right)
 {
