+++ rtgui/tools/defringe.cc
@@ -16,16 +16,17 @@
  *  You should have received a copy of the GNU General Public License
  *  along with RawTherapee.  If not, see <https://www.gnu.org/licenses/>.
  */
-#include <cmath>
-#include <iomanip>
-
 #include "defringe.h"
-#include "curveeditor.h"
-#include "curveeditorgroup.h"
+
 #include "options.h"
+#include "widgets/curves/curveeditor.h"
+#include "widgets/curves/curveeditorgroup.h"
 
 #include "rtengine/color.h"
 #include "rtengine/procparams.h"
+
+#include <cmath>
+#include <iomanip>
 
 using namespace rtengine;
 using namespace rtengine::procparams;
