+++ rtgui/tools/icmpanel.cc
@@ -20,21 +20,22 @@
 
 #include "icmpanel.h"
 
+#include "editcallbacks.h"
 #include "eventmapper.h"
 #include "guiutils.h"
 #include "labgrid.h"
 #include "options.h"
 #include "pathutils.h"
 #include "rtimage.h"
-#include "curveeditor.h"
-#include "curveeditorgroup.h"
-#include "editcallbacks.h"
-#include <unistd.h>
+#include "widgets/curves/curveeditor.h"
+#include "widgets/curves/curveeditorgroup.h"
 
 #include "rtengine/dcp.h"
 #include "rtengine/iccstore.h"
 #include "rtengine/procparams.h"
 #include "rtengine/utils.h"
+
+#include <unistd.h>
 
 using namespace rtengine;
 using namespace rtengine::procparams;
