+++ rtgui/tools/locallabtools.h
@@ -21,12 +21,12 @@
 #ifndef _LOCALLABTOOLS_H_
 #define _LOCALLABTOOLS_H_
 
-#include "curveeditorgroup.h"
-#include "curveeditor.h"
+#include "adjuster.h"
 #include "labgrid.h"
 #include "thresholdadjuster.h"
 #include "toolpanel.h"
-#include "adjuster.h"
+#include "widgets/curves/curveeditorgroup.h"
+#include "widgets/curves/curveeditor.h"
 
 /* ==== LocallabToolListener ==== */
 class LocallabTool;
