+++ rtgui/tools/rgbcurves.cc
@@ -18,9 +18,9 @@
  */
 #include "rgbcurves.h"
 
-#include "curveeditor.h"
-#include "curveeditorgroup.h"
 #include "options.h"
+#include "widgets/curves/curveeditor.h"
+#include "widgets/curves/curveeditorgroup.h"
 
 #include "rtengine/procparams.h"
