+++ rtgui/tools/vibrance.cc
@@ -19,9 +19,10 @@
 
 #include "vibrance.h"
 
-#include "curveeditor.h"
-#include "curveeditorgroup.h"
 #include "options.h"
+#include "widgets/curves/curveeditor.h"
+#include "widgets/curves/curveeditorgroup.h"
+
 #include "rtengine/color.h"
 
 using namespace rtengine;
