+++ rtgui/widgets/curves/curveeditor.cc
