+++ rtgui/widgets/curves/curveeditor.h
