+++ rtgui/widgets/curves/curveeditorgroup.cc
