+++ rtgui/widgets/curves/curveeditorgroup.h
