+++ rtgui/widgets/curves/diagonalcurveeditorsubgroup.cc
