+++ rtgui/widgets/curves/diagonalcurveeditorsubgroup.h
