+++ rtgui/widgets/curves/flatcurveeditorsubgroup.cc
