+++ rtgui/widgets/curves/flatcurveeditorsubgroup.h
