+++ rtgui/widgets/curves/mycurve.cc
