+++ rtgui/widgets/curves/mycurve.h
