+++ rtgui/widgets/curves/mydiagonalcurve.cc
