+++ rtgui/widgets/curves/mydiagonalcurve.h
