+++ rtgui/widgets/curves/myflatcurve.cc
