+++ rtgui/widgets/curves/myflatcurve.h
